refactor: new file structure
