+++ packages/openapi-parser/src/utils/openapi/actions/get.ts
@@ -1,5 +1,5 @@
-import type { CommandChain, Queue, Task } from '../../types'
-import { workThroughQueue } from './utils/workThroughQueue'
+import type { CommandChain, Queue, Task } from '../../../types'
+import { workThroughQueue } from '../utils/workThroughQueue'
 
 /**
  * Run the chained tasks and return the results
