+++ packages/openapi-parser/src/utils/openapi/actions/toYaml.ts
@@ -1,7 +1,7 @@
-import type { Queue, Task } from '../../types'
-import { getEntrypoint } from '../getEntrypoint'
-import { toYaml as toYamlUtility } from '../toYaml'
-import { workThroughQueue } from './utils/workThroughQueue'
+import type { Queue, Task } from '../../../types'
+import { getEntrypoint } from '../../getEntrypoint'
+import { toYaml as toYamlUtility } from '../../toYaml'
+import { workThroughQueue } from '../utils/workThroughQueue'
 
 /**
  * Run the chained tasks and return the results
