+++ packages/openapi-parser/src/utils/openapi/commands/dereferenceCommand.ts
@@ -1,11 +1,11 @@
-import type { DereferenceResult, Queue, Task } from '../../types'
-import type { DereferenceOptions } from '../dereference'
-import { details } from './details'
-import { files } from './files'
-import { get } from './get'
-import { toJson } from './toJson'
-import { toYaml } from './toYaml'
-import { queueTask } from './utils/queueTask'
+import type { DereferenceResult, Queue, Task } from '../../../types'
+import type { DereferenceOptions } from '../../dereference'
+import { details } from '../actions/details'
+import { files } from '../actions/files'
+import { get } from '../actions/get'
+import { toJson } from '../actions/toJson'
+import { toYaml } from '../actions/toYaml'
+import { queueTask } from '../utils/queueTask'
 
 declare global {
   // eslint-disable-next-line @typescript-eslint/consistent-type-definitions
