+++ packages/openapi-parser/src/utils/openapi/commands/loadCommand.ts
@@ -4,19 +4,19 @@ import type {
   LoadResult,
   Queue,
   Task,
-} from '../../types'
-import type { DereferenceOptions } from '../dereference'
-import type { LoadOptions } from '../load'
-import type { ValidateOptions } from '../validate'
+} from '../../../types'
+import type { DereferenceOptions } from '../../dereference'
+import type { LoadOptions } from '../../load'
+import type { ValidateOptions } from '../../validate'
+import { details } from '../actions/details'
+import { files } from '../actions/files'
+import { get } from '../actions/get'
+import { toJson } from '../actions/toJson'
+import { toYaml } from '../actions/toYaml'
+import { queueTask } from '../utils/queueTask'
 import { dereferenceCommand } from './dereferenceCommand'
-import { details } from './details'
-import { files } from './files'
 import { filterCommand } from './filterCommand'
-import { get } from './get'
-import { toJson } from './toJson'
-import { toYaml } from './toYaml'
 import { upgradeCommand } from './upgradeCommand'
-import { queueTask } from './utils/queueTask'
 import { validateCommand } from './validateCommand'
 
 declare global {
