+++ packages/openapi-parser/src/utils/openapi/commands/upgradeCommand.ts
@@ -1,20 +1,14 @@
-import type {
-  AnyApiDefinitionFormat,
-  AnyObject,
-  Queue,
-  Task,
-  UpgradeResult,
-} from '../../types'
-import type { DereferenceOptions } from '../dereference'
-import type { ValidateOptions } from '../validate'
+import type { AnyObject, Queue, Task, UpgradeResult } from '../../../types'
+import type { DereferenceOptions } from '../../dereference'
+import type { ValidateOptions } from '../../validate'
+import { details } from '../actions/details'
+import { files } from '../actions/files'
+import { get } from '../actions/get'
+import { toJson } from '../actions/toJson'
+import { toYaml } from '../actions/toYaml'
+import { queueTask } from '../utils/queueTask'
 import { dereferenceCommand } from './dereferenceCommand'
-import { details } from './details'
-import { files } from './files'
 import { filterCommand } from './filterCommand'
-import { get } from './get'
-import { toJson } from './toJson'
-import { toYaml } from './toYaml'
-import { queueTask } from './utils/queueTask'
 import { validateCommand } from './validateCommand'
 
 declare global {
