+++ packages/openapi-parser/src/utils/openapi/commands/validateCommand.ts
@@ -1,21 +1,15 @@
-import type {
-  AnyApiDefinitionFormat,
-  AnyObject,
-  Queue,
-  Task,
-  ValidateResult,
-} from '../../types'
-import type { DereferenceOptions } from '../dereference'
-import type { ValidateOptions } from '../validate'
+import type { AnyObject, Queue, Task, ValidateResult } from '../../../types'
+import type { DereferenceOptions } from '../../dereference'
+import type { ValidateOptions } from '../../validate'
+import { details } from '../actions/details'
+import { files } from '../actions/files'
+import { get } from '../actions/get'
+import { toJson } from '../actions/toJson'
+import { toYaml } from '../actions/toYaml'
+import { queueTask } from '../utils/queueTask'
 import { dereferenceCommand } from './dereferenceCommand'
-import { details } from './details'
-import { files } from './files'
 import { filterCommand } from './filterCommand'
-import { get } from './get'
-import { toJson } from './toJson'
-import { toYaml } from './toYaml'
 import { upgradeCommand } from './upgradeCommand'
-import { queueTask } from './utils/queueTask'
 
 declare global {
   // eslint-disable-next-line @typescript-eslint/consistent-type-definitions
